[dominion] revert Hero
