+++ dominion/dominion_ui/src/components/Hero/HomeHero.js
@@ -7,8 +7,16 @@ import { withStyles } from '@material-ui/core/styles';
 import herobg from '../../assets/images/bg/hero_bg.png';
 import arrow from '../../assets/images/icons/combined-shape.svg';
 import map from '../../assets/images/bg/hero_map.png';
+import background from '../../assets/images/bg/background.png';
 
 const styles = theme => ({
+  root: {
+    flexGrow: 1,
+    padding: theme.spacing.unit * 4,
+    backgroundImage: `url(${background})`,
+    backgroundRepeat: 'no-repeat',
+    backgroundSize: 'cover'
+  },
   heroContentGrid: {
     flexGrow: 1,
     backgroundImage: `url(${herobg})`,
@@ -83,68 +91,82 @@ const styles = theme => ({
   }
 });
 
-function Hero({ classes }) {
+function HomeHero({ classes }) {
   return (
-    <Grid
-      container
-      direction="row"
-      spacing={40}
-      justify="space-between"
-      className={classes.heroContentGrid}
-    >
-      <Grid item xs={8} md={8} lg={8} xl={6} className={classes.titleTextGrid}>
-        <Typography variant="h1" className={classes.text}>
-          Discover the stories behind the data.
-        </Typography>
-
-        <Typography item xs={4} variant="body1" className={classes.body2}>
-          Dominion makes data available to help add context and authority to
-          public discourse and policy-making on vital issues of land ownership.
-        </Typography>
-        <Grid style={{ paddingTop: '2rem' }}>
-          <a href="/#" className={classes.buttonLink}>
-            <Button variant="outlined" color="white" className={classes.button}>
-              Select a Country
-            </Button>
-            <img
-              src={arrow}
-              alt="Select Country"
-              className={classes.buttonArrow}
-            />
-          </a>
-        </Grid>
-      </Grid>
+    <Grid container item xs={12} className={classes.root}>
       <Grid
-        xs={4}
-        md={4}
-        lg={4}
-        xl={5}
-        item
         container
         direction="row"
-        justify="space-around"
-        alignItems="center"
-        className={classes.mapSection}
+        spacing={40}
+        justify="space-between"
+        className={classes.heroContentGrid}
       >
-        <img src={map} alt="Country Map" />
-        <Typography
-          variant="fontSmallDefault"
-          className={classes.verticalAlignText}
+        <Grid
+          item
+          xs={8}
+          md={8}
+          lg={8}
+          xl={6}
+          className={classes.titleTextGrid}
         >
-          South Africa &nbsp;&nbsp;
-          <div
-            style={{ display: 'inline-block', borderLeft: '1px solid #fff' }}
+          <Typography variant="h1" className={classes.text}>
+            Discover the stories behind the data.
+          </Typography>
+
+          <Typography item xs={4} variant="body1" className={classes.body2}>
+            Dominion makes data available to help add context and authority to
+            public discourse and policy-making on vital issues of land
+            ownership.
+          </Typography>
+          <Grid style={{ paddingTop: '2rem' }}>
+            <a href="/#" className={classes.buttonLink}>
+              <Button
+                variant="outlined"
+                color="white"
+                className={classes.button}
+              >
+                Select a Country
+              </Button>
+              <img
+                src={arrow}
+                alt="Select Country"
+                className={classes.buttonArrow}
+              />
+            </a>
+          </Grid>
+        </Grid>
+        <Grid
+          xs={4}
+          md={4}
+          lg={4}
+          xl={5}
+          item
+          container
+          direction="row"
+          justify="space-around"
+          alignItems="center"
+          className={classes.mapSection}
+        >
+          <img src={map} alt="Country Map" />
+          <Typography
+            variant="fontSmallDefault"
+            className={classes.verticalAlignText}
           >
-            &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;{' '}
-          </div>
-        </Typography>
+            South Africa &nbsp;&nbsp;
+            <div
+              style={{ display: 'inline-block', borderLeft: '1px solid #fff' }}
+            >
+              &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;{' '}
+            </div>
+          </Typography>
+        </Grid>
       </Grid>
     </Grid>
   );
 }
 
-Hero.propTypes = {
+HomeHero.propTypes = {
   classes: PropTypes.isRequired
 };
 
-export default withStyles(styles)(Hero);
+export default withStyles(styles)(HomeHero);
